requirements.txt
numpy==1.23.3
pandas==1.4.2
plotly==5.4.0
requests==2.28.1
streamlit
web3==6.0.0
autopep8
pre-commit
python-dotenv
pytest